chore: 쿼리 hook으로 분리
